DynamicDo_Flask/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.8.3
click==8.1.8
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
openai==2.0.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.1
python-dotenv==1.1.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
